requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0
scikit-image>=0.21.0
scipy>=1.11.0
PyYAML>=6.0
tqdm>=4.65.0
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pytest>=7.4.0
kaggle>=1.5.16
